Major refactor of search
