+++ core/mongo/__init__.py
@@ -1,0 +1,1 @@
+from sanic_motor import BaseModel
