+++ ons/search/conceptual/search_engine.py
@@ -1,70 +1,67 @@
-from core.search.search_engine import SearchEngine
+from ons.search.search_engine import SearchEngine
+from ons.search.sort_fields import SortFields
+
+from numpy import ndarray
 
 
 class ConceptualSearchEngine(SearchEngine):
     from server.word_embedding.sanic_supervised_models import load_model
     from core.word_embedding.models.supervised import SupervisedModels, SupervisedModel
 
+    user_vector_key = "user_vector"
     word_embedding_model: SupervisedModel = load_model(SupervisedModels.ONS)
-
-    def __init__(self, **kwargs):
-        super(ConceptualSearchEngine, self).__init__(**kwargs)
 
     def content_query(
             self,
             search_term: str,
             current_page: int = 1,
             size: int = 10,
+            sort_by: SortFields=SortFields.relevance,
             **kwargs):
         """
         Overwrite SearchEngine content query to use a vector rescore.
         :param search_term:
         :param current_page:
         :param size:
+        :param sort_by:
         :param kwargs:
         :return:
         """
-        from core.search.fields import embedding_vector
-        from core.search.sort_by import SortFields
-        from core.search.conceptual_search.conceptual_search_queries import content_query
-
-        from numpy import ndarray
-
-        kwargs_copy = kwargs.copy()
-        sort_by = kwargs_copy.pop("sort_by", SortFields.relevance)
+        from ons.search.fields import embedding_vector
+        from ons.search.sort_fields import SortFields
+        from ons.search.conceptual.queries import content_query
 
         if sort_by == SortFields.relevance:
             # Build the content query with vector function score
             query = content_query(
                 search_term,
                 ConceptualSearchEngine.word_embedding_model,
-                **kwargs_copy)
+                **kwargs)
 
-            # Prepare the final query and omit the embedding_vector field from
-            # _source
-            query_dict = {
-                "query": query.to_dict()
-            }
+            # Build the query
+            s: ConceptualSearchEngine = self.build(
+                query,
+                search_term=search_term,
+                current_page=current_page,
+                size=size,
+                sort_by=None,  # None to prevent sorting from influencing scores
+                **kwargs)
 
+            # Add the rescore?
             # If user_vector is specified, add a user vector function score
-            if 'user_vector' in kwargs:
-                from core.search.conceptual_search.conceptual_search_queries import user_rescore_query
-                user_vector: ndarray = kwargs.get('user_vector')
+            if self.user_vector_key in kwargs:
+                from ons.search.conceptual.queries import user_rescore_query
+
+                user_vector: ndarray = kwargs.get(self.user_vector_key)
 
                 if user_vector is not None and isinstance(
                         user_vector, ndarray):
 
-                    query_dict["rescore"] = user_rescore_query(user_vector)
-
-            s = self.build_query(
-                query_dict,
-                search_term=search_term,
-                current_page=current_page,
-                size=size,
-                **kwargs_copy)
+                    rescore = user_rescore_query(user_vector)
+                    s: ConceptualSearchEngine = s.extra(**rescore.to_dict())
 
             # Exclude embedding vector for source
-            s = s.source(exclude=[embedding_vector.name])
+            s: ConceptualSearchEngine = s.source(exclude=[embedding_vector.name])
 
             return s
         else:
@@ -82,7 +79,7 @@ class ConceptualSearchEngine(SearchEngine):
         :param search_term:
         :return:
         """
-        from core.search.content_types import home_page_census, product_page
+        from ons.search.content_type import home_page_census, product_page
 
         type_filters = [product_page.name, home_page_census.name]
