+++ server/search/routes.py
@@ -6,7 +6,7 @@ from sanic.exceptions import InvalidUsage
 from server.search import execute_search
 from server.requests import get_form_param
 
-from core.search.search_engine import SearchEngine
+from ons.search.search_engine import SearchEngine
 
 search_blueprint = Blueprint('search', url_prefix='/search')
 
@@ -18,7 +18,7 @@ async def search(request: Request) -> HTTPResponse:
     :param request:
     :return:
     """
-    from core.search.filter_functions import all_filter_funcs
+    from ons.search.filter_functions import all_filter_funcs
 
     search_term = request.args.get("q")
     if search_term is not None:
@@ -37,7 +37,7 @@ async def search_data(request: Request) -> HTTPResponse:
     :param request:
     :return:
     """
-    from core.search.filter_functions import filters
+    from ons.search.filter_functions import filters
 
     search_term = request.args.get("q")
     if search_term is not None:
@@ -56,7 +56,7 @@ async def search_publications(request: Request) -> HTTPResponse:
     :param request:
     :return:
     """
-    from core.search.filter_functions import filters
+    from ons.search.filter_functions import filters
 
     search_term = request.args.get("q")
     if search_term is not None:
@@ -75,8 +75,8 @@ async def search_departments(request: Request) -> HTTPResponse:
     :param request:
     :return:
     """
-    from core.search.indices import Index
-    from core.search.response import marshall_hits
+    from ons.search.indicies import Index
+    from ons.search.response import marshall_hits
 
     search_term = request.args.get("q")
     if search_term is not None:
@@ -116,10 +116,10 @@ async def find_document_by_uri(request: Request, path: str='') -> HTTPResponse:
     :param path:
     :return:
     """
-    from core.search.indices import Index
-    from core.search.response import marshall_hits
-    from core.search.queries import match_by_uri
-    from core.search.search_engine import SearchEngine
+    from ons.search.indicies import Index
+    from ons.search.response import marshall_hits
+    from ons.search.queries import match_by_uri
+    from ons.search.search_engine import SearchEngine
 
     from sanic.exceptions import NotFound
